# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank cards API
  description: REST API Documentation
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /users:
    post:
      tags:
        - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /users/{id}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    patch:
      tags:
        - user-controller
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /users/all:
    get:
      tags:
        - user-controller
      operationId: getAllUsers
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
        - name: userFilter
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UserFilter"
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedResponseUserResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "404 NOT_FOUND"
        message:
          type: string
          example: "User not found"
    CreateUserRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          example: "89991234455"
        password:
          type: string
          maxLength: 20
          minLength: 8
          example: "cool-password"
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
      required:
        - password
        - phoneNumber
        - role
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        phoneNumber:
          type: string
          example: "89991234455"
        role:
          type: string
          example: "ROLE_ADMIN"
...
    UpdateUserRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          example: "89991234455"
        password:
          type: string
          maxLength: 20
          minLength: 8
          example: "cool-password"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
          default: 0
        size:
          type: integer
          format: int32
          minimum: 1
          default: 30
        sort:
          type: array
          items:
            type: string
    UserFilter:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        phoneNumber:
          type: string
          example: 89991234455
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
    PagedResponseUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        size:
          type: integer
          format: int32
          example: 30
        page:
          type: integer
          format: int32
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 33
        totalPages:
          type: integer
          format: int32
          example: 2
        last:
          type: boolean
          example: false
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT